backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.11.7
motor
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2
